backend/requirements.txt
Django==3.2.3
django-filter==23.2
django-model-utils==4.5.1
djangorestframework==3.12.4
python-dotenv==1.0.0
django-extensions==3.2.3
flake8==6.1.0
flake8-django==1.4